.gitignore

+++ conf/evolutions/default/1.sql
@@ -4,7 +4,7 @@
 # --- !Ups
 
 create table solucao_do_exercicio (
-  id                        bigint not null,
+  id                        integer not null,
   codigo                    varchar(255),
   constraint pk_solucao_do_exercicio primary key (id))
 ;
